# public/openapi.yaml
openapi: 3.0.3
info:
  title: PHP API Starter
  version: 1.0.0
  description: Simple no-framework PHP API with routing, middleware, and OAuth2 auth.
servers:
  - url: /pub-api/public

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  time:
                    type: string
                    format: date-time
  /test:
    get:
      summary: Test endpoint
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: string
  /v1/users:
    get:
      summary: List users
      security:
        - OAuth2accident: ["accident-api"]
        - ApiKeyAuth: ["accident-api"]

      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
    post:
      summary: Create user
      security:
        - OAuth2accident: ["accident-api"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation error
  /v1/users/{id}:
    get:
      summary: Get user
      security:
        - OAuth2accident: ["accident-api"]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not found

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    OAuth2accident:
      type: oauth2
      description: IdentityServer OAuth2 for accident SPA (Authorization Code with PKCE)
      flows:
        authorizationCode:
          authorizationUrl: https://localhost:5001/connect/authorize
          tokenUrl: https://localhost:5001/connect/token
          scopes:
            accident-api: Access accident API
            openid: OpenID Connect
            profile: User profile

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string

security:
  - OAuth2accident: ["accident-api"]